+++ tools/evm-toolkit/src/lib/common/interfaces/index.ts
