+++ tools/evm-toolkit/src/lib/common/utils.ts
