+++ tools/evm-toolkit/src/lib/iscmagic/index.ts
