+++ tools/evm-toolkit/src/lib/iscmagic/iscmagic.ts
