+++ tools/evm-toolkit/src/lib/iscmagic/types/web3-v1-contracts/types.ts
